Concurrency. Download the file.
Downloads the csv file and saves to disk. I need to work on both my TDD
approach, smaller steps to speed me up, and the quality of the tests I
am writing. Using dependency injection and mocking a File well would
allow me to assert that the contents written to the file are correct.

But, bored of this right now so moving on.